+++ desktop/cmp/dash/report/DashReport.js
@@ -2,18 +2,18 @@ import {ContextMenu} from '@blueprintjs/core';
 import classNames from 'classnames';
 import {div} from '@xh/hoist/cmp/layout';
 import {elemFactory, hoistCmp, uses, XH} from '@xh/hoist/core';
-import {dashGridLayoutView} from './impl/DashGridLayoutView';
-import {dashGridLayoutContainerContextMenu} from './impl/DashGridLayoutContainerContextMenu';
-import {DashGridLayoutContainerModel} from './DashGridLayoutContainerModel';
-import './DashGridLayoutContainer.scss';
+import {dashReportView} from './impl/DashReportView';
+import {dashReportContextMenu} from './impl/DashReportContextMenu';
+import {DashReportModel} from './DashReportModel';
+import './DashReport.scss';
 
 import ReactGridLayout, {WidthProvider} from 'react-grid-layout';
 import 'react-grid-layout/css/styles.css';
 const reactGridLayout = elemFactory(WidthProvider(ReactGridLayout));
 
-export const dashGridLayoutContainer = hoistCmp.factory({
-    className: 'dash-grid-layout-container',
-    model: uses(DashGridLayoutContainerModel),
+export const dashReport = hoistCmp.factory({
+    className: 'xh-dash-report',
+    model: uses(DashReportModel),
     render({className, model}) {
         return div({
             className: classNames(
@@ -25,7 +25,7 @@ export const dashGridLayoutContainer = hoistCmp.factory({
             onContextMenu: (e) => {
                 const {clientX: x, clientY: y} = e;
                 ContextMenu.show(
-                    dashGridLayoutContainerContextMenu({dashGridLayoutContainerModel: model,
+                    dashReportContextMenu({dashGridLayoutContainerModel: model,
                         clickPosition: {x, y}}),
                     {left: x, top: y},
                     null,
@@ -50,7 +50,7 @@ export const dashGridLayoutContainer = hoistCmp.factory({
                 onLayoutChange: (layout) => model.setLayout(layout),
                 items: model.viewModels.map(vm => div({
                     key: vm.id,
-                    item: dashGridLayoutView({model: vm})
+                    item: dashReportView({model: vm})
                 }))
             })
         });
