+++ desktop/cmp/dash/report/DashReport.scss
@@ -1,4 +1,4 @@
-.dash-grid-layout-container {
+.xh-dash-report {
   width: 100%;
   height: 100%;
   overflow: auto;
