+++ desktop/cmp/dash/report/DashReportViewModel.js
@@ -7,9 +7,9 @@
 import {DashViewModel} from '@xh/hoist/desktop/cmp/dash/DashViewModel';
 
 /**
- * Model for a content item within a DashGridLayoutContainer. Extends {@see DashViewModel}
+ * Model for a content item within a DashReport. Extends {@see DashViewModel}
  */
-export class DashGridLayoutViewModel extends DashViewModel {
+export class DashReportViewModel extends DashViewModel {
     get positionParams() {
         const {containerModel, id} = this;
         return containerModel.layout.find(view => view.i === id);
