--- desktop/cmp/dashGrid/index.js
@@ -1,4 +1,0 @@
-export * from './DashGridLayoutContainer';
-export * from './DashGridLayoutContainerModel';
-export * from './DashGridLayoutViewModel';
-export * from './DashGridLayoutViewSpec';
